Repair dataflow_utils null comparison
We updated code in dataflow_utils to account for strings in the
ChangeDataTypes function being passed into math.isnan() function to
prevent crashes. The math.isnan() function only works with numeric
entities, and will raise an exception when a string is passed into it.
This fix allows for the police_blotter_30_day DAG to run properly.