Additional repair for dataflow_utils processing
The previous update failed to handle nan entities properly, as they would
continue to be evaluated as integers, strings, floats, etc. even after
their values were set to None. This update adds a condition to skip past
the data update process if an entity were already determined to have a
value of None.