Fix domi_permits geocode_address function call
Previously, the computronix_domi_permits_dataflow script would fail due to
a TypeError caused by an incorrect number of parameters being passed to
the dataflow_utils.geocode_address function. This commit passes the
correct number of parameters to geocode_address so that the DAG will not
fail when run on our local machines.