Change datatype refactor (#103)
* Repair dataflow_utils null comparison

We updated code in dataflow_utils to account for strings in the
ChangeDataTypes function being passed into math.isnan() function to
prevent crashes. The math.isnan() function only works with numeric
entities, and will raise an exception when a string is passed into it.
This fix allows for the police_blotter_30_day DAG to run properly.

* Additional repair for dataflow_utils processing

The previous update failed to handle nan entities properly, as they would
continue to be evaluated as integers, strings, floats, etc. even after
their values were set to None. This update adds a condition to skip past
the data update process if an entity were already determined to have a
value of None.

Co-authored-by: Jason Ficorilli <ip.analytics@pittsburghpa.gov>