Dataflow test refactor (#104)
* Repair dataflow_utils null comparison

We updated code in dataflow_utils to account for strings in the
ChangeDataTypes function being passed into math.isnan() function to
prevent crashes. The math.isnan() function only works with numeric
entities, and will raise an exception when a string is passed into it.
This fix allows for the police_blotter_30_day DAG to run properly.

* Modify test for dataflow_utils ChangeDataTypes

We updated the ChangeDataTypes testing function to ensure that all
variable types were processed correctly. One large issue was that NaN
values were first changed to None and then, as the next code block was
executed, None was changed to a string or incompatible data type. The
solution was to skip the second set of conversions. We then ensured that
all data types were accounted for in the test by setting None values to be
transformed to floats, ints, and strings. The test verified that if the
values were nan then that conversion would be skipped and the value would
instead be changed to None.

Co-authored-by: Jason Ficorilli <ip.analytics@pittsburghpa.gov>
Co-authored-by: woodjt <37873653+woodjt@users.noreply.github.com>