Fix Unix Timestamp to Datetime String Function
This commit fixes an error in the unix_to_date_string dataflow_util
function, which previously took in a Unix timestamp and returned a UTC
date string that was incorrectly labeled to be in Eastern time format.
The updated function returns a UTC and Eastern time datetime string with
correct formatting.