Small changes to dataflow utils standardizeTimes function
The return type of the dataflow utils standardizeTimes function was UTC, EST and Unix time stamps but we returned datetime objects till now. Made changes to the return type so that we now have the following:
{
  'UTC': string,
  'EST', string,
  'Unix': float
}
Also changed the TZ_CONVERTER dictionary declaration to a more concise representation using the dict() constructor instead of a literal.