Fix Department Names Test Function
This commit adjusts the input to the StandardizeDepNames testing function
so that the expected result matches the actual result. This required a
typo to be fixed in the list of regex strings, and also required the
output of the utility function to be stored in a DataFrame so that the
expected value does not get compared to a generator variable.