Remove Timezone Input Hashing
This commit removes the TZ_CONVERTER dictionary from the StandardizeTimes
dataflow util function. Now instead of accepting faulty input from the
user and correcting it afterward, the function will expect the user to
pass an acceptable timezone name into the function. If the timezone they
name is not accepted by pytz, the function will now pass a ValueError.