Modified the current test case for standardize time dataflow util function
We enhanced the current function which tests the dataflow_utils standardize time. Initially we were hard coding the time difference between UTC and Eastern zones which would fail once the daylight savings time kicks in/out in the fall/spring months. Hence we make use of the inbuilt          utc_offset() function in the datetime library which can easily identify whether to account for daylights savings or not. This ensures that the test cases run for any given timestamp.