Add time conversion dataflow_util (#112)
* Add time conversion dataflow_util

This commit adds a new class to dataflow_utils that takes in datetime
string values and standardizes them to datetimes in UTC, EST, and Unix
formats. It is powerful enough to handle datetimes in a variety of
timezones and string formats.

* Fix Unix Timestamp to Datetime String Function

This commit fixes an error in the unix_to_date_string dataflow_util
function, which previously took in a Unix timestamp and returned a UTC
date string that was incorrectly labeled to be in Eastern time format.
The updated function returns a UTC and Eastern time datetime string with
correct formatting.

* Small changes to dataflow utils standardizeTimes function

The return type of the dataflow utils standardizeTimes function was UTC, EST and Unix time stamps but we returned datetime objects till now. Made changes to the return type so that we now have the following:
{
  'UTC': string,
  'EST', string,
  'Unix': float
}
Also changed the TZ_CONVERTER dictionary declaration to a more concise representation using the dict() constructor instead of a literal.

* Extend the standardizeTime test case to include API calls

The current unit test function for the standardizeTime dataflow utils class uses a parser to parse hard coded datetime strings. This additional test case has a dynamic datetime input with different formats for each run cycle. We make an API call to fetch the Unix timestamp, UTC and EST times. The API response is then verified against the values retrieved from the dataflow utils time standardize function.

* Remove Timezone Input Hashing

This commit removes the TZ_CONVERTER dictionary from the StandardizeTimes
dataflow util function. Now instead of accepting faulty input from the
user and correcting it afterward, the function will expect the user to
pass an acceptable timezone name into the function. If the timezone they
name is not accepted by pytz, the function will now pass a ValueError.

* Modified the current test case for standardize time dataflow util function

We enhanced the current function which tests the dataflow_utils standardize time. Initially we were hard coding the time difference between UTC and Eastern zones which would fail once the daylight savings time kicks in/out in the fall/spring months. Hence we make use of the inbuilt          utc_offset() function in the datetime library which can easily identify whether to account for daylights savings or not. This ensures that the test cases run for any given timestamp.

* Rename EST to EAST in Timezone Standardization to Remove Ambiguity

* Update StandardizeTimes DocStrings and Randomize Timedelta for Test Function

Co-authored-by: Jason Ficorilli <ip.analytics@pittsburghpa.gov>
Co-authored-by: Pranav Banthia <pranavbanthia28@gmail.com>