Changes to test standardize time function
The current version of the standardize_times testing function made use of API calls to convert a given timestamp into unix, eastern and UTC timestamps. This function had a flaw when it came to processing timestamps which were in and around the daylight savings time period. Current changes corrected those flaws and the test function now processes randomly selected timestamps with random formats for all of pytz's common timezones (around 500 timezones on that list).