Added a utility function to mask VIP names
The wireless dashboards project had data from Verizon, Sprint and AT&T which includes city employees who have been assigned a device handset with a number. Certain members from this list belong to VIP's and their information needs to be masked before processing and releasing any PII on the dashboard. Hence we write a utility function with the aim of anonymizing this information.