Refactor Datestring From Unix Utility Function
This commit ensures continuity from the soon-to-be-deprecated GetDateStrings
dataflow utility function to the updated version, which returns datestrings
in UTC and Eastern timezone formats. The two functions now exist in the code
next to each other, which allows us to comment out the original function
when it is time to switch to the new one. In a previous branch, the old
function was entirely overwritten by the new one. This commit fixes that.