Add Field to GeoWrapper Output to Mark Intersections
This commit adds a new field to the datums returned by the GeoWrapper
dataflow utility function that are marked either True or False depending
on whether a given address corresponds to an intersection or not. It also
returns datums with placeholder fields with null values if an address is
determined to be invalid in some way. For example, an imprecise address is
given a lat, long value of None, None since no precise coordinates can be
determined.