Add Single Address Type Field, Merge 2 Geo Utils
This commit alters the output of the GeoWrapper dataflow class to make use
of a single field to mark the status of an address: address_status. This
field will designate an address as a precise address, an intersection, an
underspecified address, or an invalid address. Additionally, the
functionality of regularize_address has been merged into geocode_address,
which reduces the number of functions called by the GeoWrapper. The
geocode_address function now makes use of the Google Maps API instead of
the Allegheny County API due to its ability to find lat/long values for
street intersections.