Changes to test standardize time function (#118)
* Changes to test standardize time function

The current version of the standardize_times testing function made use of API calls to convert a given timestamp into unix, eastern and UTC timestamps. This function had a flaw when it came to processing timestamps which were in and around the daylight savings time period. Current changes corrected those flaws and the test function now processes randomly selected timestamps with random formats for all of pytz's common timezones (around 500 timezones on that list).

* Clean up comments in standardize_times test func

The previous commit had an inaccurate comment regarding the inner-workings of the standardize_times testing function, as well as a lingering print statement that was present for debugging purposes. This commit cleans up the test function code so that it can be deployed to production.

Co-authored-by: Pranav Banthia <pranavbanthia28@gmail.com>
Co-authored-by: jasonfic <43254577+jasonfic@users.noreply.github.com>