Add Google Address Column Alongside User-Specified One
This commit adds a dependency to the address formatting and geocoding
portion of the GeoWrapper Beam utility function that will preserve the
address information as entered by a 311 requester, while also including
a fully formatted address returned by the Google Maps API. The new
procedure is to retain all information returned by the GeoWrapper, but
depend on the address_type flag to decide which information to include or
ignore. For example, an address marked as Unmappable will have its invalid
Google formatted address name available in the BQ table, but an analyst
will be able to filter out Unmappable addresses and hide it from a dashboard.