Control for Empty Addresses With & Without Coords
This commit adds new dependencies to the id_underspecified_addresses util
function to account for addresses that were left blank by the user. If
these addresses contain lat/long values, they are identified as Coordinates
Only in the address_type field. If there is no street name or lat/long
provided, then the address is identified as Missing. These types of
addresses are not passed on to be formatted or geocoded.