Rename Function, Add Test Cases, Fix Date Error
This commit renames the find_last_updated_file airflow utility function to
find_backfill_date in order to more effectively communicate the purpose of
the function. It also adds 3 more test cases to the airflow test function
in order to prove its effectiveness across Google Cloud projects and
buckets. Finally, the backfill function no longer has an error in which
the upload dates for the files in a bucket were off by one day. This issue
was occurring due to a failure to convert to local time from UTC.