Determine backfill date (#123)
* initial commit of pulled branch

* initial commit of pulled branch

* rewrite of date check

* this and the previous commit made 2 changes- we changed the function name; we also changed the way the last run was extracted. the blob creation date was listed in UTC (despite being EST in the GCP console). we needed to add a conversion to EST and the problem was solved.

* Rename Function, Add Test Cases, Fix Date Error

This commit renames the find_last_updated_file airflow utility function to
find_backfill_date in order to more effectively communicate the purpose of
the function. It also adds 3 more test cases to the airflow test function
in order to prove its effectiveness across Google Cloud projects and
buckets. Finally, the backfill function no longer has an error in which
the upload dates for the files in a bucket were off by one day. This issue
was occurring due to a failure to convert to local time from UTC.

* Add Dependency for an Empty or Null Bucket

* Change Variable From Reserved Word, Add Comments

Co-authored-by: Jesse Wood <woodjt.mail@gmail.com>
Co-authored-by: Jason Ficorilli <ip.analytics@pittsburghpa.gov>