Change Accepted Geo Function Params to a List
This commit edits the args accepted by the GoogleMapsClassifyAndGeocode
utility function to take in a single list of strings instead of several
column names separated by commas. Additionally, all references to the
within_city_bounds helper function have been removed, since that step will
now be performed with a BigQuery operator when the DAG is executed.