Create Airflow Util to Mark Addresses Outside City
This commit adds a function to airflow_utils called within_city_bounds,
which takes in the lat and long of an address and uses them in a SQL query
to determine if they fall within the borders of the City of Pittsburgh,
while not falling within the borders of Mt. Oliver. If an address is
within the city, it is returned as-is. Otherwise, it will be flagged as
Outside of City in the address_type column.