Changes to Reverse Geo Query to include DPW start and end date
DPW (Dept of public works) changes its boundaries for streets, divisions and
parks. We recently got the data for all the DPW boundaries both current and
back dated. We now use the boundary creation date as a filter when trying to
build a geo query. The 311 ticket creation date is checked against the DPW
boundary creation and end date to find out which division, street and park the
ticket belonged to when it was created.