Use data-rivers As Source Project for SELECT Query
This commit rectifies a mistake in the previous commit, in which an
environmental variable is to obtain the Google Cloud project name for the
qalert.requests table. Since this table only exists in data-rivers and not
the test project environment, I removed the environmental variable and
hardcoded in the proper Google Cloud Project name.