Add Test Functions, Fix Var Name & Comment Typos
This commit introduces a couple of new tests for the dataflow utility
functions developed for the updated Computronix pipeline (ConvertBooleans
and FilterOutliers). Over the course of the design of the test functions,
a handful of syntax errors were discovered. There were a handful of minor
typos in the comments, and a few variables pointed to the wrong location.
Adjustments were made to make the utility functions run properly.