JSON Linter Function (#137)
* Json Linter function for ndjson files

Sometimes ndjson files have two json objects in the same line, hence it
is necessary to separate such json objects into individual lines. We parse
each line and check sequentially for such occurrence.

* Add Test Function for Json Linter Util (Incomplete)

* Adjust JSON Linter to Handle 2+ JSONs in 1 Line

This commit adds a chunk of code to the json_linter GCS Loader utility
function that accounts for a source json that has a delimiter error in
which multiple JSON records are present on the same line, when a new line
is expected as a delimiter. The json_linter test function is also adjusted
to ensure that this new functionality works properly.

* Add Test Case with Multiple Rows

Co-authored-by: Pranav Banthia <pranavbanthia28@gmail.com>
Co-authored-by: Jason Ficorilli <ip.analytics@pittsburghpa.gov>