Add New Sub-Folder to Qalert Bucket for Lat Longs
Currently, Qalert backfill DAGs decide which day to start backfilling from
based on a json file in the successful_run_log folder. Without creating a
new sub-folder to host a separate successful_run_log file, this targetted
backfill focused just on lat longs will be incomplete because it will only
run up to the date of the last regular backfill (sometime in March). With
a fresh sub-folder, the first run for this DAG will start from the very
first set of Qalert records instead of only starting in March.