Add All Stages of Electric Car Charging Session DAG
This commit provides the Airflow and Dataflow stages for the Chargepoint
DAG. The Dataflow tasks are lightweight formatting changes, such as
renaming and dropping columns and providing time data in a variety of time
zones. The Airflow stage provides locations for storage of JSON and AVRO
files in GCS from which Dataflow transformations and BigQuery storage are
sourced from. Finally, a Format/Deduplication query is provided which will
remove duplicate charging session records and set columns in the proper
order.