Merge Chargepoint DAG into Data-Rivers-Testing Environment (#164)
* Migrate Chargepoint Development into Branch with Up-To-Date Utility Functions (#163)

* Init GCS Loader Stage of New DAG for Electric Vehicle Charging Data

* Change Backfill JSON File Naming Convention to Use Date Instead of Record Number

* Upload All Data as Single JSON, Let DAG Dynamically Decide How Much Data to Pull

* Upload All Progress on Chargepoint DAG and Merge With Data-Rivers-Testing Branch

Co-authored-by: Jason Ficorilli <ip.analytics@pittsburghpa.gov>

* Add All Stages of Electric Car Charging Session DAG

This commit provides the Airflow and Dataflow stages for the Chargepoint
DAG. The Dataflow tasks are lightweight formatting changes, such as
renaming and dropping columns and providing time data in a variety of time
zones. The Airflow stage provides locations for storage of JSON and AVRO
files in GCS from which Dataflow transformations and BigQuery storage are
sourced from. Finally, a Format/Deduplication query is provided which will
remove duplicate charging session records and set columns in the proper
order.

* Remove Successful Run Logging from Loop, Make Small Change to Airflow Comment

* Remove Unused Variable

Co-authored-by: Jason Ficorilli <ip.analytics@pittsburghpa.gov>