Prevent KeyErrors in SwapFieldNames Util Function (#267)
The InTime DAG experienced a dataflow failure stemming from the
SwapFieldNames utility function. After some investigation, it turned out
that KeyErrors occur within the function if a field whose name we are
trying to swap is missing from a row of the datum. This change addresses
this issue by handling the exception, printing a message to the console,
and setting the value of the new field to None.

Co-authored-by: Jason Ficorilli <ip.analytics@pittsburghpa.gov>