Data rivers testing (#270)
* Prevent KeyErrors in SwapFieldNames Util Function (#267)

The InTime DAG experienced a dataflow failure stemming from the
SwapFieldNames utility function. After some investigation, it turned out
that KeyErrors occur within the function if a field whose name we are
trying to swap is missing from a row of the datum. This change addresses
this issue by handling the exception, printing a message to the console,
and setting the value of the new field to None.

Co-authored-by: Jason Ficorilli <ip.analytics@pittsburghpa.gov>

* Change Subdirectory Name to prev_day_weather for Improved Clarity (#268)

Co-authored-by: Jason Ficorilli <ip.analytics@pittsburghpa.gov>

* Remove Redundancy in Weather Avro Path (#269)

Co-authored-by: Jason Ficorilli <ip.analytics@pittsburghpa.gov>

Co-authored-by: jasonfic <43254577+jasonfic@users.noreply.github.com>
Co-authored-by: Jason Ficorilli <ip.analytics@pittsburghpa.gov>