Geocode Unmapped Cartegraph Tasks (#273)
* Prevent KeyErrors in SwapFieldNames Util Function (#267)

The InTime DAG experienced a dataflow failure stemming from the
SwapFieldNames utility function. After some investigation, it turned out
that KeyErrors occur within the function if a field whose name we are
trying to swap is missing from a row of the datum. This change addresses
this issue by handling the exception, printing a message to the console,
and setting the value of the new field to None.

Co-authored-by: Jason Ficorilli <ip.analytics@pittsburghpa.gov>

* Change Subdirectory Name to prev_day_weather for Improved Clarity (#268)

Co-authored-by: Jason Ficorilli <ip.analytics@pittsburghpa.gov>

* Create Small Cleanup DAG to Add Geo-Zones to Unmapped Cartegraph Tasks

* Add Airflow Util that Splits the Timebound Rev-Geo Query into Individual Pieces

* Initialize Temporary Table at Beginning of DAG, Merge at End

* Add Some Whitespace

* Format Temp Table So it Matches All Tasks Table and Can Be Joined with UNION ALL

* Shorten Task Names, Adjust Join Tables Query, Get Rid of Duplicates Caused by Points Falling on Geo-Borders

Co-authored-by: Jason Ficorilli <ip.analytics@pittsburghpa.gov>