Delete SendGrid Import from Dataflow Utils
There is currently a DAG failure occurring in some pipelines stemming from
a ModuleNotFoundError when attempting to import sendgrid. This error is
only occurring in the dataflow_utils script and not elsewhere where the
library is imported. Because data quality checking now resides in
airflow_utils and not dataflow_utils, there is no reason to keep the
sendgrid import in dataflow, especially if it is causing crashes.